refactored LandingPage
