+++ src/Components/LandingPage/LandingPage.css
@@ -1,33 +1,57 @@
 .Header {
-    min-height: 80px;
+  /* background-image: url("../images/maxresdefault.jpg"); */
+  background-repeat: no-repeat;
+  color: #5b7bc2;
+  background-color: rgba(86, 21, 170, 0.842);
+  min-height: 80px;
     max-height: 80px;
   }
+  
   .Header .navbar {
     max-width: 1500px;
     margin-left: auto;
     margin-right: auto;
+    color: #5b7bc2;
   }
   .Header .SpaceRace {
-    margin-top: 1rem;
-    width: 100%;
+    position:center;
+    margin-top: 2rem;
+    padding: right 22px;
+    color: #5b7bc2;
+
   }
   .Header .SpaceRace .name {
-    margin-bottom: 4px;
-    font-size: 1.5rem;
+    margin: 22px;
+    font-size: 2.5rem;
     font-weight: bold;
-    color: #3b5998;
+    color: #5b7bc2;
   }
   .Header .SpaceRace .title {
-    color: #2eadf7;
+    color: #41b0f0;
+    margin: 10px;
+    font-size: 3.5rem;
   }
   
+body {
+  /* background-image: url("../images/maxresdefault.jpg"); */
+  background-size: cover;
 
+  background-image: url('../images/maxresdefault.jpg');
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-color: rgb(1, 1, 7); 
+  padding: 10px;
+  font-size: 2.0rem;
+  position: relative;
+  margin: 2
+}
 @import './Font/font.css';
 .hero-image {
   /* The image used */
-  background-image: url("../images/cockpit.png");
+  background-image: url("./cockpit6.png");
   /* Set a specific height */
-  height: 300px;
+  height: 550px;
+  
   /* Position and center the image to scale nicely on all screens */
   background-position: center;
   background-repeat: no-repeat;
@@ -42,16 +66,22 @@
   position: absolute;
   top: 10%;
   left: 50%;
+  
+  /* text-size-adjust: 30px; */
   transform: translate(-50%, -50%);
   color: #3f52f8;
 }
+
 .title {
   font-family: 'zigzageoregular';
   text-align: center;
   text-shadow: -0.5px -0.5px 0 #eea928, -1px -1px 0 #f3d964, -3px -2px 0 #6b3ee7, -3px -3px 0 #eb1430, 1px 1px 0 #eea928;
   position: absolute;
+  margin:  top 3rem;
+  
   top: 45%;
   left: 50%;
+
   transform: translate(-50%, -50%);
   color: #3f52f8;
 }
@@ -59,7 +89,8 @@
     animation-delay: 2s;
     padding-top: 0;
     padding-left: 2%;
-    background-color: #333;
+    font-size: 2rem;
+    background-color: rgba(86, 21, 170, 0.842);
   }
   .Footer .footer-nav {
     flex-direction: column;
@@ -170,19 +201,19 @@
   }
   @media (min-width: 736px) {
     .Footer .wrapper {
-      margin-top: 0;
+      margin-top: -0.44;
     }
   }
   .Footer .wrapper .left {
     float: left;
   }
   .Footer .wrapper .left .name {
-    font-size: 0.9rem;
-    line-height: 0.2rem;
+    font-size: 1.0rem;
+    line-height: 0.75rem;
   }
   @media (min-width: 992px) {
     .Footer .wrapper .left .name {
-      font-size: 1.1rem;
+      font-size: 2.1rem;
     }
   }
   .Footer .wrapper .right {
@@ -196,5 +227,14 @@
     .Footer .wrapper .right .domain {
       font-size: 1.1rem;
     }
+  }
+  .stripe {
+    background-image: url('./stripe.png');
+    background-position: center;
+  background-repeat: no-repeat;
+  background-size: contain;
+  width: 33px;
+  height: 33px;
+  position: relative;
   }
