+++ src/Components/LandingPage/Middle.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Button } from 'reactstrap';
-import './middle.scss'
+// import './middle.scss'
 /****
  * 
  * So far this just has a background image with title and start button
@@ -18,12 +18,13 @@ class Middle extends Component {
                 <div className="hero-image"> 
                     <img  />                 
                         <div className="hero-text">
-                            <Button><b>START</b></Button>
+                            <Button style={{width: '155px', height: '40px', color: 'red', backgroundColor: '#5b7bc2'}}><h3><b>START</b></h3></Button>
                             <br/>
                             <br/>
                         </div>
                         <div className="title">
-                            <h1><strong><b>SPACE  RACE</b></strong></h1>
+                           
+                            
                         </div>
                 </div>
             </div>
