--- src/Components/LandingPage/middle.scss
@@ -1,47 +1,0 @@
-@import './Font/font.css';
-.hero-image {
-   
-    /* The image used */
-    background-image: url("../images/cockpit.png");
-  
-    /* Set a specific height */ 
-   height: 300px;
- 
-   /* Position and center the image to scale nicely on all screens */
-   background-position: center;
-   background-repeat: no-repeat;
-   background-size: contain;
-   position: relative; 
- }
- 
- /* Place text in the middle of the image */
- .hero-text {
-   font-family: 'zigzageoregular';
-   text-align: center;
-   text-shadow:
-   -.5px -.5px 0 rgb(238, 169, 40),
-  -1px -1px 0 rgb(243, 217, 100),  
-   -3px -2px 0 rgb(107, 62, 231),
-  -3px -3px 0 rgb(235, 20, 48),
-   1px 1px 0 rgb(238, 169, 40);
-   position: absolute;
-   top: 10%;
-   left: 50%;
-   transform: translate(-50%, -50%);
-   color: rgb(63, 82, 248);
- }
- .title {
-   font-family: 'zigzageoregular';
-   text-align: center;
-   text-shadow:
-   -.5px -.5px 0 rgb(238, 169, 40),
-  -1px -1px 0 rgb(243, 217, 100),  
-   -3px -2px 0 rgb(107, 62, 231),
-  -3px -3px 0 rgb(235, 20, 48),
-   1px 1px 0 rgb(238, 169, 40);
-   position: absolute;
-   top: 45%;
-   left: 50%;
-   transform: translate(-50%, -50%);
-   color: rgb(63, 82, 248);
- }
